* New Action/Factory for loading reach attributes. * TestHelper now has methods for loading/writing serialized objects for easier benchmark object state comparison. * RM unused sample files.
Former-commit-id: 1515ee7103998e60348ce0054fb7998dbd9eed3e

+++ src/main/java/gov/usgswim/sparrow/cachefactory/ReachByIDFactory.java
@@ -39,7 +39,7 @@ public class ReachByIDFactory extends AbstractCacheFactory {
 
 		String query = getTextWithParamSubstitution(
 				"FindReach",
-				new String[] { "ModelId", Long.toString(req.getModelID()), "Identifier", Integer.toString( req.getReachID() ) });
+				new String[] { "ModelId", Long.toString(req.getModelID()), "Identifier", Long.toString( req.getReachID() ) });
 		System.out.println("SPRW reachById: " + query);
